requirements.txt
pypdf 
sentence_transformers 
Pinecone
chromadb
pymilvus 
google-generativeai
dotenv